React doesn't like when you add styles into pages Moved all styles to their own styles folder
