+++ package.json
@@ -19,6 +19,7 @@
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "axios": "^0.27.2",
+    "eslint-config-react-app": "^7.0.1",
     "hamburger-react": "^2.5.0",
     "moment": "^2.29.4",
     "next": "^12.3.1",
