+++ src/components/hackathon-list.js
@@ -1,5 +1,4 @@
 import React from "react";
-import { useAuth0 } from "@auth0/auth0-react";
 import EventFeature from "./event-feature";
 import useHackathonEvents from '../hooks/use-hackathon-events';
